requirements.txt
asgiref==3.11.1
diff-match-patch==20241021
Django==6.0.4
django-import-export==4.4.0
django-rest-framework==0.1.0
djangorestframework==3.17.1
et_xmlfile==2.0.0
openpyxl==3.1.5
pillow==12.2.0
sqlparse==0.5.5
tablib==3.9.0
tzdata==2026.1
xlrd==2.0.2
xlwt==1.3.0
